+++ tests/unit/test_convert_json_to_junit_xml.py
@@ -5,7 +5,7 @@ import xml.etree.ElementTree as ET
 import pytest
 from pathlib import Path
 
-from pytest_reporter_plus.json_to_xml_converter import convert_json_to_junit_xml, sanitize_classname
+from pytest_html_plus.json_to_xml_converter import convert_json_to_junit_xml, sanitize_classname
 
 
 @pytest.mark.skip(reason="Skipping for test coverage")
